fix(smi): make index iteration agree with keys()
SIM118 wanted `for k in d.keys()` written as `for k in d`, which broke MIB
label resolution: sysDescr came back under joint-iso-itu-t instead of iso.

The cause is in pysnmp/smi/indices.py. OrderedDict and OidOrderedDict exist
to impose an order that is not insertion order, and they override keys(),
values() and items() to give it -- but not __iter__. So `for k in d`,
`list(d)`, `dict(d)` and `**d` all fell through to dict's insertion order and
quietly returned a different sequence from the one keys() returns. Every
caller that happened to drop a `.keys()` got the wrong answer.

Override __iter__ to return the same order, and pin it with tests: keys that
sort differently as strings than as OIDs, walked every way.

Also fixes a `NotImplementedError()` that was constructed and discarded
rather than raised, and two `raise exc_v` inside the handler that caught it.

Refs #178.

+++ pysnmp/proto/secmod/eso/priv/aesbase.py
@@ -31,7 +31,7 @@ class AbstractAesBlumenthal(aes.Aes):
         localPrivKey = localkey.localizeKey(privKey, snmpEngineID, hashAlgo)
 
         # now extend this key if too short by repeating steps that includes the hashPassphrase step
-        for count in range(1, int(ceil(self.keySize * 1.0 / len(localPrivKey)))):
+        for count in range(1, ceil(self.keySize * 1.0 / len(localPrivKey))):
             localPrivKey += localPrivKey.clone(
                 hashAlgo(localPrivKey.asOctets()).digest()
             )
